# Fix performance regression by updating findfirst calls for 0.7