--- a/REQUIRE
+++ b/REQUIRE
@@ -2,4 +2,4 @@ julia 0.6
 IntervalSets 0.1
 IterTools
 RangeArrays
-Compat 0.33.0
+Compat 0.61.0