--- a/src/indexing.jl
+++ b/src/indexing.jl
@@ -284,14 +284,14 @@ end
 
 # Categorical axes may be indexed by their elements
 function axisindexes(::Type{Categorical}, ax::AbstractVector, idx)
-    i = findfirst(ax, idx)
-    i == 0 && throw(ArgumentError("index $idx not found"))
+    i = Compat.findfirst(isequal(idx), ax)
+    i === nothing && throw(ArgumentError("index $idx not found"))
     i
 end
 function axisindexes(::Type{Categorical}, ax::AbstractVector, idx::Value)
     val = idx.val
-    i = findfirst(ax, val)
-    i == 0 && throw(ArgumentError("index $val not found"))
+    i = Compat.findfirst(isequal(val), ax)
+    i === nothing && throw(ArgumentError("index $val not found"))
     i
 end
 # Categorical axes may be indexed by a vector of their elements