Fixes #136 

## On v0.7

Before: `214.838 μs (210 allocations: 24.53 KiB)`
After: `24.563 ns (1 allocation: 16 bytes)`

## On v0.6

Before: `49.175 ns (1 allocation: 16 bytes)`
After: `42.095 ns (3 allocations: 48 bytes)`